chore: split operations into their individual uses
