+++ src/enums.ts
@@ -10,10 +10,3 @@ export enum PRStatus {
   IN_FLIGHT = 'in-flight/',
   NEEDS_MANUAL = 'needs-manual-bp/',
 }
-
-// trop repo setup constants
-export enum TropAction {
-  INIT_REPO = 'INIT_REPO',
-  SET_UP_REMOTES = 'SET_UP_REMOTES',
-  BACKPORT = 'BACKPORT',
-}
